pyproject.toml
[project]
name = "upscaler"
version = "0.1.0"
description = "AI-powered image upscaler using Real-ESRGAN with web API"
requires-python = ">=3.12"
dependencies = [
    "basicsr>=1.4.2",
    "fastapi>=0.121.1",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "realesrgan>=0.3.0",
    "torch<2.5",
    "torchvision<0.20",
    "uvicorn[standard]>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
    "black>=24.0.0",
    "httpx>=0.27.0",
]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "N812"]  # Line too long (handled by black), Allow non-lowercase imports

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[dependency-groups]
dev = [
    "black",
    "httpx",
    "pytest",
    "pytest-cov",
    "ruff",
]

